backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
langchain>=0.1.16
langchain-core>=0.1.52
langchain-openai>=0.0.8
langchain-community>=0.0.38
langgraph>=0.0.26
weaviate-client==4.4.1
tavily-python==0.3.0
pydantic>=2.6.0
aiohttp==3.9.3
sse-starlette==1.8.2
python-multipart==0.0.6
